# Double registers per iteration in FP32-arithmetic FP16 ARM gemv kernel